Added NutritionalValueService and ShoppingListService tests

+++ Services/CookWithMe.Services.Data/NutritionalValues/NutritionalValueService.cs
@@ -1,5 +1,6 @@
 ﻿namespace CookWithMe.Services.Data.NutritionalValues
 {
+    using System;
     using System.Threading.Tasks;
 
     using CookWithMe.Data.Common.Repositories;
@@ -11,6 +12,9 @@
 
     public class NutritionalValueService : INutritionalValueService
     {
+        private const string InvalidNutritionalValueIdErrorMessage = "NutritionalValue with ID: {0} does not exist.";
+        private const string InvalidRecipeIdErrorMessage = "NutritionalValue with RecipeId: {0} does not exist.";
+
         private readonly IDeletableEntityRepository<NutritionalValue> nutritionalValueRepository;
 
         public NutritionalValueService(IDeletableEntityRepository<NutritionalValue> nutritionalValueRepository)
@@ -23,6 +27,12 @@
             var nutritionalValue = await this.nutritionalValueRepository
                 .GetByIdWithDeletedAsync(id);
 
+            if (nutritionalValue == null)
+            {
+                throw new ArgumentNullException(
+                    string.Format(InvalidNutritionalValueIdErrorMessage, id));
+            }
+
             this.nutritionalValueRepository.Delete(nutritionalValue);
             var result = await this.nutritionalValueRepository.SaveChangesAsync();
 
@@ -33,6 +43,12 @@
         {
             var nutritionalValueFromDb = await this.nutritionalValueRepository
                 .GetByIdWithDeletedAsync(id);
+
+            if (nutritionalValueFromDb == null)
+            {
+                throw new ArgumentNullException(
+                    string.Format(InvalidNutritionalValueIdErrorMessage, id));
+            }
 
             nutritionalValueFromDb.Calories = nutritionalValueServiceModel.Calories;
             nutritionalValueFromDb.Fats = nutritionalValueServiceModel.Fats;
@@ -51,6 +67,12 @@
             var nutritionalValue = await this.nutritionalValueRepository
                 .GetByIdWithDeletedAsync(id);
 
+            if (nutritionalValue == null)
+            {
+                throw new ArgumentNullException(
+                    string.Format(InvalidNutritionalValueIdErrorMessage, id));
+            }
+
             return nutritionalValue.To<NutritionalValueServiceModel>();
         }
 
@@ -59,6 +81,12 @@
             var nutritionalValue = await this.nutritionalValueRepository
                 .AllAsNoTracking()
                 .SingleOrDefaultAsync(x => x.RecipeId == recipeId);
+
+            if (nutritionalValue == null)
+            {
+                throw new ArgumentNullException(
+                    string.Format(InvalidRecipeIdErrorMessage, recipeId));
+            }
 
             return nutritionalValue.Id;
         }
